Project discontinued. (click to read more)
Changelog:
- Shortened the code for the money decimals display.
- Pressing backspace while in battle would open up the help page and the default key bind wouldn't work.
- Changed the updater's domain url request, since the htssoft.tk domain died.
- Balanced some material drops.

Discontinued Project:
Unfortunately, HTS-Soft was shut down for the team's personal reasons. This game is "temporally discontinued" - which means that the next update will be massively delayed. We're sorry.

On the other hand, there is now a "remastered" version of this game in the works by someone else. The read me file will be edited to redirect to that project when it's fully out.