The Combat Update #3.25
Every time a single enemy dies, it will display an animation on that single enemy. The animation used to play on all enemies at the same time once the battle ended and the player had won.
Updated the script command, use `script /?` on the terminal to learn more.
Made sure that all enemies look towards the player.